requirements.txt
astroid==2.3.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
py-getch==1.0.1
pylint==2.4.4
pyserial==3.4
rope==0.14.0
six==1.13.0
typed-ast==1.4.0
wrapt==1.11.2
